+++ tests/Integration/DSC_SqlRS.config.ps1
@@ -17,7 +17,13 @@ else
     $mockLastDrive = ((Get-Volume).DriveLetter | Sort-Object | Select-Object -Last 1)
     $mockIsoMediaDriveLetter = [char](([int][char]$mockLastDrive) + 1)
 
-    if($script:sqlVersion -eq '140')
+    if($script:sqlVersion -eq '150')
+    {
+        # SQL2019
+        $instanceName = 'SSRS'
+        $isoImageName = 'SQL2019.iso'
+    }
+    elseif($script:sqlVersion -eq '140')
     {
         # SQL2017
         $instanceName = 'SSRS'
@@ -74,20 +80,6 @@ Configuration DSC_SqlRS_CreateDependencies_Config
 
     node $AllNodes.NodeName
     {
-        MountImage 'MountIsoMedia'
-        {
-            ImagePath   = $Node.ImagePath
-            DriveLetter = $Node.DriveLetter
-            Ensure      = 'Present'
-        }
-
-        WaitForVolume 'WaitForMountOfIsoMedia'
-        {
-            DriveLetter      = $Node.DriveLetter
-            RetryIntervalSec = 5
-            RetryCount       = 10
-        }
-
         User 'CreateReportingServicesServiceAccount'
         {
             Ensure   = 'Present'
@@ -105,6 +97,20 @@ Configuration DSC_SqlRS_CreateDependencies_Config
 
         if($script:sqlVersion -eq '130')
         {
+            MountImage 'MountIsoMedia'
+            {
+                ImagePath   = $Node.ImagePath
+                DriveLetter = $Node.DriveLetter
+                Ensure      = 'Present'
+            }
+
+            WaitForVolume 'WaitForMountOfIsoMedia'
+            {
+                DriveLetter      = $Node.DriveLetter
+                RetryIntervalSec = 5
+                RetryCount       = 10
+            }
+
             SqlSetup 'InstallReportingServicesInstance'
             {
                 InstanceName          = $Node.InstanceName
@@ -133,10 +139,10 @@ Configuration DSC_SqlRS_CreateDependencies_Config
             }
         }
         <#
-            DSC_SqlRSSetup.Integration.Tests.ps1 will have installed SSRS 2017.
+            DSC_SqlRSSetup.Integration.Tests.ps1 will have installed SSRS 2017 or 2019.
             We just need to start SSRS.
         #>
-        elseif($script:sqlVersion -eq '140')
+        elseif($script:sqlVersion -in @('150','140'))
         {
             Service 'StartReportingServicesInstance'
             {
@@ -257,7 +263,7 @@ Configuration DSC_SqlRS_StopReportingServicesInstance_Config
                 State = 'Stopped'
             }
         }
-        elseif($script:sqlVersion -eq '140')
+        elseif($script:sqlVersion -in @('150','140'))
         {
             Service 'StopReportingServicesInstance'
             {